+++ src/concurrently.ts
@@ -1,21 +1,34 @@
 import { array } from './array';
 import { isPromise } from './is_promise';
-import { NArray, NFn } from './utilities';
+import { If, NArray, NFn } from './utilities';
+
+type ConcurrentlyParameters<V extends ReadonlyArray<any>> = NArray.Flatten<
+  NArray.TrimLastEmpty<NFn.ParametersOf<V>>
+>;
 
 /** Executes functions simultaneously and can return arrays of execution results. */
 export const concurrently =
   <F extends ReadonlyArray<(...args: ReadonlyArray<any>) => any>>(...fns: F) =>
   async (
-    ...args: NArray.Flatten<NArray.TrimLastEmpty<NFn.ParametersOf<F>>>
-  ): Promise<NFn.ReturnTypesOf<F>> =>
-    Promise.all(
+    ...args: If<
+      NArray.IsSameInnerType<ConcurrentlyParameters<F>>,
+      ConcurrentlyParameters<F> | [NArray.First<ConcurrentlyParameters<F>>],
+      ConcurrentlyParameters<F>
+    >
+  ): Promise<NFn.ReturnTypesOf<F>> => {
+    const parameters =
+      (args as any[]).length === 0
+        ? fns.map(() => [])
+        : (args as any[]).length === 1
+        ? fns.map(() => array((args as any[])[0]))
+        : (args as any[]).map((value) => array(value));
+
+    return Promise.all(
       fns.map(
         (fn, index) =>
           new Promise((resolve, reject) => {
-            const parameters = array((args as any[])[index]);
-
             try {
-              const result = fn(...parameters);
+              const result = fn(...parameters[index]);
               isPromise(result)
                 ? result.then(resolve, reject)
                 : resolve(result);
@@ -25,3 +38,4 @@ export const concurrently =
           }),
       ),
     ) as NFn.ReturnTypesOf<F>;
+  };